+++ phpstan-bootstrap.php
@@ -11,4 +11,5 @@
  * all functions and classes automatically.
  */
 
+require_once __DIR__ . '/vendor/autoload.php';
 require_once __DIR__ . '/src/config.php';